+++ src/app/locales/en/utils/space-actions/add-harvester-to-space-action.js
@@ -3,6 +3,6 @@ export default {
   modalHeader: 'Add one of your harvesters',
   modalDescription: 'Choose harvester which should consume metadata from space "{{spaceName}}":',
   modalSubmit: 'Add',
-  successNotificationText: 'The harvester has been sucessfully added to the space.',
+  successNotificationText: 'The harvester has been successfully added to the space.',
   failureNotificationActionName: 'adding the harvester to the space',
 };