+++ src/app/locales/en/utils/space-actions/remove-space-action.js
@@ -5,6 +5,6 @@ export default {
   modalDisclaimer: '<strong>This operation is irreversible</strong>. All the data in the space will be lost, although for some storage backends it might be possible to import the storage data into another space.',
   modalCheckboxMessage: 'I understand and want to proceed.',
   modalYes: 'Remove',
-  successNotificationText: 'The space has been sucessfully removed.',
+  successNotificationText: 'The space has been successfully removed.',
   failureNotificationActionName: 'removing the space',
 };