+++ src/app/locales/en/utils/workflow-actions/remove-atm-inventory-action.js
@@ -3,6 +3,6 @@ export default {
   modalHeader: 'Remove automation inventory',
   modalDescription: 'You are about to delete the automation inventory <strong>{{atmInventoryName}}</strong>.',
   modalYes: 'Remove',
-  successNotificationText: 'The automation inventory has been sucessfully removed.',
+  successNotificationText: 'The automation inventory has been successfully removed.',
   failureNotificationActionName: 'removing the automation inventory',
 };